Aktualizacja
Wyrzucone wszystek zostało i wrzucone na nowo, ale i tak nikt tego nie
czyta (a jeśli czyta to trudno).
