+++ Conspiracy/Debug/Gra.tlog/Gra.lastbuildstate
@@ -1,2 +1,2 @@
 #TargetFrameworkVersion=v4.0:PlatformToolSet=v120:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit
-Debug|Win32|C:\Users\John\Documents\Visual Studio 2013\Projects\Nasza-gra-master\Pulpit\Conspiracy\|
+Debug|Win32|C:\Users\John\Documents\GitHub\NasaHorosaIgra\Conspiracy\|
