+++ Conspiracy/code/mechanics/Platforms.hpp
@@ -30,7 +30,7 @@ for(int i =0;i<20;i++){
 	if(f.floor[i]==1){
 			std::cout<<"Start"<<std::endl;
 		ziutek[i].setTexture(good_one(width,height));
-		ziutek[i].setColor(sf::Color(0,0,255));	
+		ziutek[i].setColor(sf::Color(0,255,255));	
 		ziutek[i].setPosition(i*width,SCRHEIGHT-height);
 		platforms.push_back(ziutek[i]);
 		std::cout<<"End"<<std::endl;
