--- Pulpit/Conspiracy/code/states/Game_Over.hpp
@@ -1,6 +1,0 @@
-class Game_Over : public Game_state {
-
-
-
-
-};
