--- Pulpit/Conspiracy/importer.hpp
@@ -1,12 +1,0 @@
-#include "code/engine/My_View.hpp"
-#include "code/engine/Primitive_Graphic_Buffer.hpp"
-#include "code/engine/World_Constants.hpp"
-
-#include "code/mechanics/Floor_Maker.hpp"
-#include "code/mechanics/Platforms.hpp"
-#include "code/mechanics/Player.hpp"
-
-#include "code/states/Intro.hpp"
-#include "code/states/Logo.hpp"
-#include "code/states/Game.hpp"
-#include "code/states/Game_Over.hpp"
